watchlist/requirements.txt
beautifulsoup4==4.8.1
blinker==1.4
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
Django==2.0
Flask==1.1.1
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mysql-connector-python==8.0.18
mysqlclient==1.4.6
numpy==1.17.4
Pillow==6.1.0
protobuf==3.11.3
pymongo==3.10.0
PyMySQL==0.9.3
pyparsing==2.4.5
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.1
redis==3.3.11
requests==2.22.0
selenium==3.141.0
six==1.13.0
soupsieve==1.9.4
SQLAlchemy==1.3.13
urllib3==1.25.3
Werkzeug==1.0.0
WTForms==2.2.1
